feat: adding eks control plane support
This is work in progress to add support for an EKS control plane. Its
enabled via a feature flag.

+++ controllers/awscluster_controller.go
@@ -24,8 +24,10 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 	infrav1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha3"
+	infrav1exp "sigs.k8s.io/cluster-api-provider-aws/exp/api/v1alpha3"
 	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/scope"
 	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/services/ec2"
 	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/services/elb"
@@ -105,23 +107,25 @@ func (r *AWSClusterReconciler) Reconcile(req ctrl.Request) (_ ctrl.Result, reter
 
 	// Handle deleted clusters
 	if !awsCluster.DeletionTimestamp.IsZero() {
-		return reconcileDelete(clusterScope)
+		return r.reconcileDelete(clusterScope)
 	}
 
 	// Handle non-deleted clusters
-	return reconcileNormal(clusterScope)
+	return r.reconcileNormal(ctx, clusterScope)
 }
 
 // TODO(ncdc): should this be a function on ClusterScope?
-func reconcileDelete(clusterScope *scope.ClusterScope) (reconcile.Result, error) {
+func (r *AWSClusterReconciler) reconcileDelete(clusterScope *scope.ClusterScope) (reconcile.Result, error) {
 	clusterScope.Info("Reconciling AWSCluster delete")
 
 	ec2svc := ec2.NewService(clusterScope)
 	elbsvc := elb.NewService(clusterScope)
 	awsCluster := clusterScope.AWSCluster
 
-	if err := elbsvc.DeleteLoadbalancers(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "error deleting load balancer for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
+	if !clusterScope.IsManagedControlPlaneOwned() {
+		if err := elbsvc.DeleteLoadbalancers(); err != nil {
+			return reconcile.Result{}, errors.Wrapf(err, "error deleting load balancer for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
+		}
 	}
 
 	if err := ec2svc.DeleteBastion(); err != nil {
@@ -139,7 +143,7 @@ func reconcileDelete(clusterScope *scope.ClusterScope) (reconcile.Result, error)
 }
 
 // TODO(ncdc): should this be a function on ClusterScope?
-func reconcileNormal(clusterScope *scope.ClusterScope) (reconcile.Result, error) {
+func (r *AWSClusterReconciler) reconcileNormal(ctx context.Context, clusterScope *scope.ClusterScope) (reconcile.Result, error) {
 	clusterScope.Info("Reconciling AWSCluster")
 
 	awsCluster := clusterScope.AWSCluster
@@ -162,37 +166,57 @@ func reconcileNormal(clusterScope *scope.ClusterScope) (reconcile.Result, error)
 		return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile bastion host for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
 	}
 
-	if err := elbService.ReconcileLoadbalancers(); err != nil {
-		return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile load balancers for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
-	}
-
-	if awsCluster.Status.Network.APIServerELB.DNSName == "" {
-		clusterScope.Info("Waiting on API server ELB DNS name")
-		return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
-	}
-
-	if _, err := net.LookupIP(awsCluster.Status.Network.APIServerELB.DNSName); err != nil {
-		clusterScope.Info("Waiting on API server ELB DNS name to resolve")
-		return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
-	}
-
-	awsCluster.Spec.ControlPlaneEndpoint = clusterv1.APIEndpoint{
-		Host: awsCluster.Status.Network.APIServerELB.DNSName,
-		Port: clusterScope.APIServerPort(),
-	}
-
-	for _, subnet := range clusterScope.Subnets().FilterPrivate() {
-		found := false
-		for _, az := range awsCluster.Status.Network.APIServerELB.AvailabilityZones {
-			if az == subnet.AvailabilityZone {
-				found = true
-				break
-			}
+	if !clusterScope.IsManagedControlPlaneOwned() {
+		if err := elbService.ReconcileLoadbalancers(); err != nil {
+			return reconcile.Result{}, errors.Wrapf(err, "failed to reconcile load balancers for AWSCluster %s/%s", awsCluster.Namespace, awsCluster.Name)
 		}
 
-		clusterScope.SetFailureDomain(subnet.AvailabilityZone, clusterv1.FailureDomainSpec{
-			ControlPlane: found,
-		})
+		if awsCluster.Status.Network.APIServerELB.DNSName == "" {
+			clusterScope.Info("Waiting on API server ELB DNS name")
+			return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
+		}
+
+		if _, err := net.LookupIP(awsCluster.Status.Network.APIServerELB.DNSName); err != nil {
+			clusterScope.Info("Waiting on API server ELB DNS name to resolve")
+			return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
+		}
+
+		awsCluster.Spec.ControlPlaneEndpoint = clusterv1.APIEndpoint{
+			Host: awsCluster.Status.Network.APIServerELB.DNSName,
+			Port: clusterScope.APIServerPort(),
+		}
+
+		for _, subnet := range clusterScope.Subnets().FilterPrivate() {
+			found := false
+			for _, az := range awsCluster.Status.Network.APIServerELB.AvailabilityZones {
+				if az == subnet.AvailabilityZone {
+					found = true
+					break
+				}
+			}
+
+			clusterScope.SetFailureDomain(subnet.AvailabilityZone, clusterv1.FailureDomainSpec{
+				ControlPlane: found,
+			})
+		}
+	} else {
+		//TODO(richardcase): get the control plane endpoint from the ManagedControlPlane
+		managedControlPlane := &infrav1exp.AWSManagedControlPlane{}
+		managedControlPlaneRef := types.NamespacedName{
+			Name:      clusterScope.Cluster.Spec.ControlPlaneRef.Name,
+			Namespace: clusterScope.Cluster.Spec.ControlPlaneRef.Namespace,
+		}
+		if err := r.Get(ctx, managedControlPlaneRef, managedControlPlane); err != nil {
+			return reconcile.Result{}, errors.Wrap(err, "failed to get managed control plane ref")
+		}
+		awsCluster.Spec.ControlPlaneEndpoint = managedControlPlane.Spec.ControlPlaneEndpoint
+
+		//TODO(richardcase): set the failure domains based on the subnets
+		for _, subnet := range clusterScope.Subnets().FilterPrivate() {
+			clusterScope.SetFailureDomain(subnet.AvailabilityZone, clusterv1.FailureDomainSpec{
+				ControlPlane: true,
+			})
+		}
 	}
 
 	awsCluster.Status.Ready = true
